# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Company API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local server"
    }
  ],
  "tags": [
    {
      "name": "Companies",
      "description": "Company endpoints"
    },
    {
      "name": "Employees",
      "description": "Employee endpoints"
    },
    {
      "name": "Categories",
      "description": "Category endpoints"
    }
  ],
  "components": {
    "responses": {
      "BadRequestError": {
        "description": "Request failed fallback",
        "content": {
          "application/json": {
            "example": {
              "error": "Request Failed",
              "message": "Unable to process request due to bad input data structure."
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "example": {
              "error": "Not found"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Data values fail validation constraints",
        "content": {
          "application/json": {
            "example": {
              "error": "Validation failed",
              "details": [
                "email must be a valid email",
                "name cannot be null"
              ]
            }
          }
        }
      },
      "ConflictError": {
        "description": "Unique data attribute conflict",
        "content": {
          "application/json": {
            "example": {
              "error": "Data conflict",
              "message": "A record with this unique attribute already exists."
            }
          }
        }
      }
    }
  },
  "paths": {
    "/employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Mousa"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "example": {
                  "total": 30,
                  "offset": 0,
                  "limit": 10,
                  "data": [
                    {
                      "id": "Xk9PZ",
                      "name": "Mousa Estefan",
                      "email": "mousa@test.com",
                      "company_id": "aBcD12"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "company_id"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Mousa Estefan"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mousa@test.com"
                  },
                  "company_id": {
                    "type": "string",
                    "description": "hashed ID",
                    "example": "aBcD12"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "Mousa Estefan",
                  "email": "mousa@test.com",
                  "company_id": "aBcD12"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee found",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "Mousa Estefan",
                  "email": "mousa@test.com",
                  "company_id": "aBcD12"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Employee"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "updated@test.com"
                  },
                  "company_id": {
                    "type": "string",
                    "description": "hashed ID",
                    "example": "aBcD12"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "Updated Employee",
                  "email": "updated@test.com",
                  "company_id": "aBcD12"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Employee deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/companies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "example": "BAE"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Filter companies by category ID",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies",
            "content": {
              "application/json": {
                "example": {
                  "total": 42,
                  "page": 1,
                  "totalPages": 5,
                  "data": [
                    {
                      "id": "Xk9PZ",
                      "name": "BAE",
                      "email": "contact@bae.com",
                      "address": "KHBP",
                      "category": [
                        {
                          "id": 1,
                          "name": "tech"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "post": {
        "tags": [
          "Companies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "SwaggerDefault"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "swagger@contact.com"
                  },
                  "address": {
                    "type": "string",
                    "example": "KHBP"
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "example": 1
                    },
                    "description": "Optional array of category IDs to link to this company"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "SwaggerDefault",
                  "email": "swagger@contact.com",
                  "address": "KHBP",
                  "category": [
                    {
                      "id": 1,
                      "name": "tech"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company found",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "BAE",
                  "email": "contact@bae.com",
                  "address": "KHBP"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "address": {
                    "type": "string"
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "example": 1
                    },
                    "description": "Replaces all linked categories with this new list"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "Xk9PZ",
                  "name": "Updated Company",
                  "email": "updated@company.com",
                  "address": "New Address"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "hashed ID"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Company deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "responses": {
          "200": {
            "description": "List of all categories",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "tech"
                  },
                  {
                    "id": 2,
                    "name": "frontend"
                  }
                ]
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "tech"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "tech"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/categories/{id}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category",
        "description": "Deletes a category and removes it from all associated companies.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    }
  }
}